Bug fix and worked on review comments
