+++ builds/jazz-build-module/service-configuration-data-loader.groovy
@@ -64,6 +64,7 @@ def loadServiceConfigurationData() {
             updateConfigValue("{casbin_type}", config_loader.ACL.DATABASE.TYPE_DB)
             updateConfigValue("{casbin_timeout}", config_loader.ACL.DATABASE.TIMEOUT)
             updateConfigValue("{inst_stack_prefix}", config_loader.INSTANCE_PREFIX)
+            updateConfigValue("{conf-region}", region)
 
             sh "sed -i -- 's/{scm_type}/${config_loader.SCM.TYPE}/g' ./config/global-config.json"
             sh "sed -i -- 's,{scm_base_url},http://${config_loader.REPOSITORY.BASE_URL},g' ./config/global-config.json"
