+++ core/jazz_acl/config/dev-config.json
@@ -5,6 +5,7 @@
   "SERVICE_USER": "{jazz_admin}",
   "TOKEN_CREDS": "{jazz_admin_creds}",
   "SERVICES_TABLE_NAME": "{inst_stack_prefix}_services_dev",
+  "REGION": "{conf-region}",
   "CASBIN": {
     "TYPE": "{casbin_type}",
     "HOST": "{casbin_host}",
