+++ core/jazz_acl/config/global-config.json
@@ -62,5 +62,18 @@
         "WRITE": 30
       }
     }
+  },
+  "ADMIN_SERVICES_POLICIES": {
+    "serviceId":"",
+    "policies":[{
+      "category": "manage",
+      "permission": "admin"
+    }, {
+      "category": "code",
+      "permission": "write"
+    }, {
+      "category": "deploy",
+      "permission": "write"
+    }]
   }
 }
